Bugfix: warnings when comparing incompatible values
If a simple comparison operator (e.g. $eq, $lt) compares a
numerical value with a non-numerical value, or vice-versa,
warnings were printed by Perl. This commit fixes that.

It also switches the project from the Artistic license to
the Apache 2.0 license.

Due to the license change, the major version number is
bumped to 3.